requirements.txt
# Telegram Bot
python-telegram-bot[job-queue]==21.3

# Web Framework
fastapi==0.111.0
uvicorn[standard]==0.30.1

# Task Queue
celery[redis]==5.4.0
redis==5.0.4

# Database
sqlalchemy==2.0.30
alembic==1.13.1
asyncpg==0.29.0
aiosqlite==0.20.0

# Git Operations
GitPython==3.1.43

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.1
pydantic==2.7.3
pydantic-settings==2.3.1

# Voice Processing (for Telegram voice messages)
openai-whisper==20231117
pydub==0.25.1

# Utilities
httpx==0.27.0
aiofiles==23.2.1
structlog==24.2.0
python-json-logger==2.0.7

# Testing
pytest==8.2.2
pytest-asyncio==0.23.7
pytest-cov==5.0.0
